refactor: split files into modules
