+++ simplemma/__init__.py
@@ -8,5 +8,7 @@ __version__ = "0.9.1"
 
 
 from .langdetect import in_target_language, lang_detector
-from .simplemma import lemmatize, lemma_iterator, text_lemmatizer, is_known
+from .lemmatizer import lemmatize, lemma_iterator, text_lemmatizer, is_known
 from .tokenizer import simple_tokenizer
+from .dictionaries import update_lang_data
+from .dictionary_pickler import *
