+++ simplemma/constants.py
@@ -1,0 +1,51 @@
+LANGLIST = [
+    "ast",
+    "bg",
+    "ca",
+    "cs",
+    "cy",
+    "da",
+    "de",
+    "el",
+    "en",
+    "enm",
+    "es",
+    "et",
+    "fa",
+    "fi",
+    "fr",
+    "ga",
+    "gd",
+    "gl",
+    "gv",
+    "hbs",
+    "hi",
+    "hu",
+    "hy",
+    "id",
+    "is",
+    "it",
+    "ka",
+    "la",
+    "lb",
+    "lt",
+    "lv",
+    "mk",
+    "ms",
+    "nb",
+    "nl",
+    "nn",
+    "pl",
+    "pt",
+    "ro",
+    "ru",
+    "se",
+    "sk",
+    "sl",
+    "sq",
+    "sv",
+    "sw",
+    "tl",
+    "tr",
+    "uk",
+]
