+++ tests/test_lemmatizer.py
@@ -1,16 +1,11 @@
 """Tests for `simplemma` package."""
 
 import logging
-import os
 import pytest
-import tempfile
 
 
 import simplemma
-from simplemma import lemmatize, lemma_iterator, simple_tokenizer, text_lemmatizer
-
-
-TEST_DIR = os.path.abspath(os.path.dirname(__file__))
+from simplemma import lemmatize
 
 logging.basicConfig(level=logging.DEBUG)
 
@@ -41,33 +36,6 @@ def test_readme():
     assert lemmatize("spaghettis", lang=("it", "fr")) == "spaghetti"
     assert lemmatize("spaghetti", lang=("it", "fr")) == "spaghetto"
     assert lemmatize("spaghettis", lang=("it", "fr"), greedy=True) == "spaghetto"
-    # tokenization and chaining
-    assert simplemma.simple_tokenizer(
-        "Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua."
-    ) == [
-        "Lorem",
-        "ipsum",
-        "dolor",
-        "sit",
-        "amet",
-        ",",
-        "consectetur",
-        "adipiscing",
-        "elit",
-        ",",
-        "sed",
-        "do",
-        "eiusmod",
-        "tempor",
-        "incididunt",
-        "ut",
-        "labore",
-        "et",
-        "dolore",
-        "magna",
-        "aliqua",
-        ".",
-    ]
     assert simplemma.text_lemmatizer(
         "Sou o intervalo entre o que desejo ser e os outros me fizeram.", lang="pt"
     ) == [
@@ -93,92 +61,62 @@ def test_readme():
 
 def test_logic():
     """Test if certain code parts correspond to the intended logic."""
-    # dict generation
-    testfile = os.path.join(TEST_DIR, "data/zz.txt")
-    # simple generation, silent mode
-    mydict = simplemma.simplemma._read_dict(testfile, "zz", silent=True)
-    assert len(mydict) == 3
-    mydict = simplemma.simplemma._load_dict(
-        "zz", listpath=os.path.join(TEST_DIR, "data"), silent=True
-    )
-    assert len(mydict) == 3
-    # log warning
-    mydict = simplemma.simplemma._read_dict(testfile, "zz", silent=False)
-    assert len(mydict) == 3
-    # different length
-    mydict = simplemma.simplemma._read_dict(testfile, "en", silent=True)
-    assert len(mydict) == 5
-    # different order
-    mydict = simplemma.simplemma._read_dict(testfile, "es", silent=True)
-    assert len(mydict) == 5
-    assert mydict["closeones"] == "closeone"
-    item = sorted(mydict.keys(), reverse=True)[0]
-    assert item == "valid-word"
-
-    # file I/O
-    assert simplemma.simplemma._determine_path("lists", "de").endswith("de.txt")
-
-    # dict pickling
-    listpath = os.path.join(TEST_DIR, "data")
-    os_handle, temp_outputfile = tempfile.mkstemp(suffix=".pkl", text=True)
-    simplemma.simplemma._pickle_dict("zz", listpath, temp_outputfile)
-
     # missing languages or faulty language codes
-    mydata = simplemma.simplemma._load_data(("de", "abc", "en"))
+    mydata = simplemma.dictionaries._load_data(("de", "abc", "en"))
     with pytest.raises(TypeError):
         lemmatize("test", lang=["test"])
     with pytest.raises(TypeError):
-        simplemma.simplemma._update_lang_data(["id", "lv"])
+        simplemma.dictionaries.update_lang_data(["id", "lv"])
 
     # searches
     with pytest.raises(TypeError):
         assert lemmatize(None, lang="en") is None
     with pytest.raises(ValueError):
         assert lemmatize("", lang="en") is None
-    assert simplemma.simplemma._suffix_search("ccc", mydata[0].dict) is None
+    assert simplemma.lemmatizer._suffix_search("ccc", mydata[0].dict) is None
 
     assert (
-        simplemma.simplemma._return_lemma("Gender-Sternchens", mydata[0].dict)
+        simplemma.lemmatizer._return_lemma("Gender-Sternchens", mydata[0].dict)
         == "Gendersternchen"
     )
     assert (
-        simplemma.simplemma._return_lemma("an-gespieltes", mydata[0].dict)
+        simplemma.lemmatizer._return_lemma("an-gespieltes", mydata[0].dict)
         == "anspielen"
     )
 
     assert (
-        simplemma.simplemma._greedy_search(
+        simplemma.lemmatizer._greedy_search(
             "getesteten", mydata[0].dict, steps=0, distance=20
         )
         == "getestet"
     )
     assert (
-        simplemma.simplemma._greedy_search(
+        simplemma.lemmatizer._greedy_search(
             "getesteten", mydata[0].dict, steps=1, distance=20
         )
         == "getestet"
     )
     assert (
-        simplemma.simplemma._greedy_search(
+        simplemma.lemmatizer._greedy_search(
             "getesteten", mydata[0].dict, steps=2, distance=20
         )
         == "testen"
     )
     assert (
-        simplemma.simplemma._greedy_search(
+        simplemma.lemmatizer._greedy_search(
             "getesteten", mydata[0].dict, steps=2, distance=2
         )
         == "getestet"
     )
 
     # prefixes
-    mydata = simplemma.simplemma._load_data(("de", "ru"))
+    mydata = simplemma.dictionaries._load_data(("de", "ru"))
     assert (
-        simplemma.simplemma._prefix_search("zerlemmatisiertes", "de", mydata[0].dict)
+        simplemma.lemmatizer._prefix_search("zerlemmatisiertes", "de", mydata[0].dict)
         == "zerlemmatisiert"
     )
     assert (
-        simplemma.simplemma._prefix_search("зафиксированные", "ru", mydata[1].dict)
+        simplemma.lemmatizer._prefix_search("зафиксированные", "ru", mydata[1].dict)
         == "зафиксированный"
     )
 
@@ -187,9 +125,9 @@ def test_convenience():
     """Test convenience functions."""
     # logic
     with pytest.raises(TypeError):
-        assert simplemma.simplemma.is_known(None, lang="en") is None
+        assert simplemma.lemmatizer.is_known(None, lang="en") is None
     with pytest.raises(ValueError):
-        assert simplemma.simplemma.is_known("", lang="en") is None
+        assert simplemma.lemmatizer.is_known("", lang="en") is None
     assert simplemma.is_known("FanCY", lang="en") is True
     # known words
     assert simplemma.is_known("Fancy-String", lang="en") is False
@@ -207,7 +145,7 @@ def test_convenience():
     ]
     text = "Nous déciderons une fois arrivées. Voilà."
     assert [
-        l for l in simplemma.simplemma.lemma_iterator(text, lang="fr", greedy=False)
+        l for l in simplemma.lemmatizer.lemma_iterator(text, lang="fr", greedy=False)
     ] == simplemma.text_lemmatizer(text, lang="fr", greedy=False)
     text = "Pepa e Iván son una pareja sentimental, ambos dedicados al doblaje de películas."
     assert (simplemma.text_lemmatizer(text, lang="es", greedy=False)) == [
@@ -248,23 +186,23 @@ def test_convenience():
 
 def test_search():
     """Test simple and greedy dict search."""
-    data = simplemma.simplemma._load_data(("en",))[0]
-    assert simplemma.simplemma._simple_search("ignorant", data.dict) == "ignorant"
-    assert simplemma.simplemma._simple_search("Ignorant", data.dict) == "ignorant"
+    data = simplemma.dictionaries._load_data(("en",))[0]
+    assert simplemma.lemmatizer._simple_search("ignorant", data.dict) == "ignorant"
+    assert simplemma.lemmatizer._simple_search("Ignorant", data.dict) == "ignorant"
     assert (
-        simplemma.simplemma._dehyphen("magni-ficent", data.dict, False) == "magnificent"
+        simplemma.lemmatizer._dehyphen("magni-ficent", data.dict, False) == "magnificent"
     )
-    assert simplemma.simplemma._dehyphen("magni-ficents", data.dict, False) is None
+    assert simplemma.lemmatizer._dehyphen("magni-ficents", data.dict, False) is None
     # assert simplemma.simplemma._greedy_search('Ignorance-Tests', datadict) == 'Ignorance-Test'
     # don't lemmatize numbers
-    assert simplemma.simplemma._return_lemma("01234", data.dict) == "01234"
+    assert simplemma.lemmatizer._return_lemma("01234", data.dict) == "01234"
     # initial or not
-    data = simplemma.simplemma._load_data(("de",))[0]
+    data = simplemma.dictionaries._load_data(("de",))[0]
     assert (
-        simplemma.simplemma._simple_search("Dritte", data.dict, initial=True) == "dritt"
+        simplemma.lemmatizer._simple_search("Dritte", data.dict, initial=True) == "dritt"
     )
     assert (
-        simplemma.simplemma._simple_search("Dritte", data.dict, initial=False)
+        simplemma.lemmatizer._simple_search("Dritte", data.dict, initial=False)
         == "Dritter"
     )
 
@@ -371,69 +309,3 @@ def test_subwords():
     assert lemmatize("zerlemmatisiertes", lang="de") == "zerlemmatisiert"
     assert lemmatize("фиксированные", lang="ru") == "фиксированный"
     assert lemmatize("зафиксированные", lang="ru") == "зафиксированный"
-
-
-def test_tokenizer():
-    # tokenization and chaining
-    text = "Sent1. Sent2\r\nSent3"
-    assert simplemma.simple_tokenizer(text) == ["Sent1", ".", "Sent2", "Sent3"]
-    assert simplemma.simple_tokenizer(text) == [
-        m[0] for m in simplemma.simple_tokenizer(text, iterate=True)
-    ]
-    assert simplemma.simple_tokenizer(
-        "200er-Inzidenz 1.000er-Inzidenz 5%-Hürde 5-%-Hürde FFP2-Masken St.-Martini-Gemeinde, Lebens-, Liebes- und Arbeitsbedingungen"
-    ) == [
-        "200er-Inzidenz",
-        "1.000er-Inzidenz",
-        "5%-Hürde",
-        "5-%-Hürde",
-        "FFP2-Masken",
-        "St.-Martini-Gemeinde",
-        ",",
-        "Lebens-",
-        ",",
-        "Liebes-",
-        "und",
-        "Arbeitsbedingungen",
-    ]
-    assert simplemma.simple_tokenizer(
-        "360-Grad-Panorama @sebastiankurz 2,5-Zimmer-Wohnung 1,2-butylketoaldehyde"
-    ) == [
-        "360-Grad-Panorama",
-        "@sebastiankurz",
-        "2,5-Zimmer-Wohnung",
-        "1,2-butylketoaldehyde",
-    ]
-    assert simplemma.simple_tokenizer(
-        "Covid-19, Covid19, Covid-19-Pandemie https://example.org/covid-test"
-    ) == [
-        "Covid-19",
-        ",",
-        "Covid19",
-        ",",
-        "Covid-19-Pandemie",
-        "https://example.org/covid-test",
-    ]
-    assert simplemma.simple_tokenizer(
-        "Test 4:1-Auswärtssieg 2,5€ €3.5 $3.5 §52, for $5, 3/5 -1.4"
-    ) == [
-        "Test",
-        "4:1-Auswärtssieg",
-        "2,5€",
-        "€3.5",
-        "$3.5",
-        "§52",
-        ",",
-        "for",
-        "$5",
-        ",",
-        "3/5",
-        "-1.4",
-    ]
-    # problem here: WDR5-„Morgenecho“
-    assert simplemma.simple_tokenizer("WDR5-„Morgenecho“") == [
-        "WDR5-",
-        "„",
-        "Morgenecho",
-        "“",
-    ]
